requirements.txt
#openai>=1.93.0
openai
python-dotenv
flask
psycopg2-binary

